restructuring code as PTP client

+++ tm4c/lib/ptp/src.c
@@ -207,7 +207,7 @@ static void doSync(PtpSource *this, PTP2_TIMESTAMP *ts) {
     uint64_t offset = (fromPtpTimestamp(ts) - tai) + this->syncDelay;
     sample->offset = (int64_t) offset;
     // record current delay
-    sample->delay = (float) (uint32_t) this->syncDelay;
+    sample->delay = (float) this->syncDelay;
     // update filter
     updateFilter(this);
 }
